#117 - Extract charts interface and charts write
